+++ src/main/java/com/example/a1project/Main.java
@@ -1,5 +1,6 @@
 package com.example.a1project;
 
+import com.example.a1project.achievement.AchievementManager;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
@@ -11,10 +12,10 @@ public class Main extends Application {
     @Override
     public void start(Stage stage) throws Exception {
         
-    	Parent root = FXMLLoader.load(getClass().getResource("Home.fxml"));
+    	Parent root = FXMLLoader.load(getClass().getResource("/fxml/Home.fxml"));
     	Scene scene = new Scene(root);
     	
-    	String css = this.getClass().getResource("Task.css").toExternalForm();
+    	String css = this.getClass().getResource("/css/Task.css").toExternalForm();
     	scene.getStylesheets().add(css);
     	
     	AchievementManager.loadAchievements();
