+++ src/main/java/com/example/a1project/achievement/Achievement.java
@@ -1,4 +1,4 @@
-package com.example.a1project;
+package com.example.a1project.achievement;
 
 public class Achievement {
