+++ src/main/java/com/example/a1project/achievement/AchievementManager.java
@@ -1,4 +1,4 @@
-package com.example.a1project;
+package com.example.a1project.achievement;
 
 import java.io.*;
 import java.util.ArrayList;
