+++ src/main/java/com/example/a1project/controllers/CardViewController.java
@@ -1,6 +1,6 @@
 package com.example.a1project.controllers;
 
-import com.example.a1project.Achievement;
+import com.example.a1project.achievement.Achievement;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
