+++ src/main/resources/fxml/MainScreen.fxml
@@ -20,7 +20,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="800.0" maxWidth="1200.0" minHeight="600.0" minWidth="1200.0" prefWidth="1200.0" stylesheets="@MainScreen.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.a1project.controllers.MainScreenController">
+<AnchorPane maxHeight="800.0" maxWidth="1200.0" minHeight="600.0" minWidth="1200.0" prefWidth="1200.0" stylesheets="@../css/MainScreen.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.a1project.controllers.MainScreenController">
    <children>
       <SplitPane dividerPositions="0.3" minWidth="1200.0" prefWidth="1200.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         <items>
@@ -112,7 +112,7 @@
                                              <children>
                                                 <ImageView fitHeight="124.0" fitWidth="131.0" pickOnBounds="true" preserveRatio="true" StackPane.alignment="CENTER_LEFT">
                                                    <image>
-                                                      <Image url="@images/pencil.png" />
+                                                      <Image url="@../images/pencil.png" />
                                                    </image>
                                                 </ImageView>
                                                 <Text fill="WHITE" text="My To Do's" StackPane.alignment="CENTER_RIGHT">
