+++ src/main/resources/fxml/Task.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="700.0" maxWidth="600.0" minHeight="700.0" minWidth="800.0" prefHeight="700.0" prefWidth="800.0" stylesheets="@Task.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.a1project.controllers.TaskController">
+<AnchorPane maxHeight="700.0" maxWidth="600.0" minHeight="700.0" minWidth="800.0" prefHeight="700.0" prefWidth="800.0" stylesheets="@../css/Task.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.a1project.controllers.TaskController">
    <children>
       <VBox layoutX="14.0" layoutY="16.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" AnchorPane.bottomAnchor="-4.0" AnchorPane.leftAnchor="14.0" AnchorPane.rightAnchor="-2.0" AnchorPane.topAnchor="16.0">
          <children>
