+++ src/main/resources/images/pencil.png
